pyproject.toml
[project]
name = "adtextract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.90.0",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.1",
    "pytesseract>=0.3.13",
]
